move strategy implementation outside Spoofax project
