+++ org.metaborg.example.externalref.app/.classpath
@@ -32,6 +32,5 @@
 			<attribute name="maven.pomderived" value="true"/>
 		</attributes>
 	</classpathentry>
-	<classpathentry kind="src" path="/org.metaborg.example.externalref.expressions"/>
 	<classpathentry kind="output" path="target/classes"/>
 </classpath>
