+++ org.metaborg.example.externalref.expressions/src/main/strategies/org/metaborg/example/externalref/expressions/strategies/InteropRegisterer.java
@@ -1,5 +1,6 @@
 package org.metaborg.example.externalref.expressions.strategies;
 
+import org.metaborg.example.externalref.context.get_external_type_0_0;
 import org.strategoxt.lang.JavaInteropRegisterer;
 import org.strategoxt.lang.Strategy;
